feat : 상품 종류 수량 API 연동 및 상태 업데이트 로직 구현
