+++ README.md
@@ -8,10 +8,10 @@
   - [x] 장바구니 아이템 삭제
 - Recoil로 상태 관리
   - [x] 상품 목록 조회
+  - [x] 상품 종류 수량
+  - [x] 개별 상품의 수량
   - [ ] 전체 상품의 선택 여부
   - [ ] 개별 상품의 선택 여부
-  - [ ] 개별 상품 종류 수량
-  - [ ] 개별 상품의 수량
   - [ ] 주문 금액
   - [ ] 배송비
   - [ ] 총 결제 금액
